New Hierophant sprites

+++ code/modules/mob/living/simple_animal/hostile/megafauna/hierophant.dm
@@ -24,7 +24,7 @@ The Hierophant's attacks are as follows, and INTENSIFY at a random chance based 
 Cross Blasts and the AoE burst gain additional range as the Hierophant loses health, while Chasers gain additional speed.
 
 When The Hierophant dies, it leaves behind its staff, which, while much weaker than when wielded by The Hierophant itself, is still quite effective.
-- The staff can place a teleport rune, allowing the user to teleport themself and their allies to the rune.
+- The staff can place a teleport beacon, allowing the user to teleport themself and their allies to the beacon.
 
 Difficulty: Hard
 
@@ -32,12 +32,11 @@ Difficulty: Hard
 
 /mob/living/simple_animal/hostile/megafauna/hierophant
 	name = "Hierophant"
-	desc = "Stolen from Hyper Light Drifter."
+	desc = "A massive machine holding an equally massive club. ."
 	health = 2500
 	maxHealth = 2500
 	attacktext = "clubs"
-	//attack_sound = 'sound/weapons/sonic_jackhammer.ogg'
-	attack_sound = "swing_hit"
+	attack_sound = 'sound/weapons/sonic_jackhammer.ogg'
 	icon_state = "hierophant"
 	icon_living = "hierophant"
 	friendly = "stares down"
@@ -45,15 +44,15 @@ Difficulty: Hard
 	faction = list("boss") //asteroid mobs? get that shit out of my beautiful square house
 	speak_emote = list("preaches")
 	armour_penetration = 50
-	melee_damage_lower = 10
-	melee_damage_upper = 10
+	melee_damage_lower = 15
+	melee_damage_upper = 15
 	speed = 1
 	move_to_delay = 10
 	ranged = 1
 	pixel_x = -16
 	ranged_cooldown_time = 40
 	aggro_vision_range = 23
-	loot = list(/obj/item/weapon/hierophant_staff)
+	loot = list(/obj/item/weapon/hierophant_club)
 	wander = FALSE
 	var/burst_range = 3 //range on burst aoe
 	var/beam_range = 5 //range on cross blast beams
@@ -61,11 +60,11 @@ Difficulty: Hard
 	var/chaser_cooldown = 101 //base cooldown/cooldown var between spawning chasers
 	var/major_attack_cooldown = 60 //base cooldown for major attacks
 	var/blinking = FALSE //if we're doing something that requires us to stand still and not attack
-	var/obj/effect/hierophant/spawned_rune //the rune we teleport back to
-	var/timeout_time = 15 //after this many Life() ticks with no target, we return to our rune
-	var/did_reset = TRUE //if we timed out, returned to our rune, and healed some
-	//var/list/kill_phrases = list("Wsyvgi sj irivkc xettih. Vitemvmrk...", "Irivkc wsyvgi jsyrh. Vitemvmrk...", "Jyip jsyrh. Egxmzexmrk vitemv gcgpiw...")
-	//var/list/target_phrases = list("Xevkix psgexih.", "Iriqc jsyrh.", "Eguymvih xevkix.")
+	var/obj/effect/hierophant/spawned_beacon //the beacon we teleport back to
+	var/timeout_time = 15 //after this many Life() ticks with no target, we return to our beacon
+	var/did_reset = TRUE //if we timed out, returned to our beacon, and healed some
+	var/list/kill_phrases = list("Wsyvgi sj irivkc xettih. Vitemvmrk...", "Irivkc wsyvgi jsyrh. Vitemvmrk...", "Jyip jsyrh. Egxmzexmrk vitemv gcgpiw...")
+	var/list/target_phrases = list("Xevkix psgexih.", "Iriqc jsyrh.", "Eguymvih xevkix.")
 	medal_type = MEDAL_PREFIX
 	score_type = BIRD_SCORE
 	del_on_death = TRUE
@@ -74,25 +73,25 @@ Difficulty: Hard
 /mob/living/simple_animal/hostile/megafauna/hierophant/New()
 	..()
 	internal = new/obj/item/device/gps/internal/hierophant(src)
-	spawned_rune = new(loc)
+	spawned_beacon = new(loc)
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/Life()
 	. = ..()
-	if(. && spawned_rune && !client)
-		if(target || loc == spawned_rune.loc)
+	if(. && spawned_beacon && !client)
+		if(target || loc == spawned_beacon.loc)
 			timeout_time = initial(timeout_time)
 		else
 			timeout_time--
 		if(timeout_time <= 0 && !did_reset)
 			did_reset = TRUE
-			//visible_message("<span class='hierophant'>\"Vixyvrmrk xs fewi...\"</span>")
-			blink(spawned_rune)
-			adjustHealth(min((health - maxHealth) * 0.5, -50)) //heal for 50% of our missing health
+			visible_message("<span class='hierophant'>\"Vixyvrmrk xs fewi...\"</span>")
+			blink(spawned_beacon)
+			adjustHealth(min((health - maxHealth) * 0.5, -250)) //heal for 50% of our missing health, minimum 10% of maximum health
 			wander = FALSE
-			/*if(health > maxHealth * 0.9)
+			if(health > maxHealth * 0.9)
 				visible_message("<span class='hierophant'>\"Vitemvw gsqtpixi. Stivexmrk ex qebmqyq ijjmgmirgc.\"</span>")
 			else
-				visible_message("<span class='hierophant'>\"Vitemvw gsqtpixi. Stivexmsrep ijjmgmirgc gsqtvsqmwih.\"</span>")*/
+				visible_message("<span class='hierophant'>\"Vitemvw gsqtpixi. Stivexmsrep ijjmgmirgc gsqtvsqmwih.\"</span>")
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/death()
 	if(health > 0 || stat == DEAD)
@@ -103,33 +102,30 @@ Difficulty: Hard
 		animate(src, alpha = 0, color = "660099", time = 20, easing = EASE_OUT)
 		addtimer(src, "update_atom_colour", 20)
 		burst_range = 10
-		//visible_message("<span class='hierophant'>\"Mrmxmexmrk wipj-hiwxvygx wiuyirgi...\"</span>")
-		visible_message("<span class='hierophant_warning'>[src] disappears in a massive burst of magic, leaving only its staff.</span>")
+		visible_message("<span class='hierophant'>\"Mrmxmexmrk wipj-hiwxvygx wiuyirgi...\"</span>")
+		visible_message("<span class='hierophant_warning'>[src] disappears in a massive burst of energy, leaving only its club.</span>")
 		burst(get_turf(src))
 		..()
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/Destroy()
-	qdel(spawned_rune)
+	qdel(spawned_beacon)
 	. = ..()
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/devour(mob/living/L)
 	for(var/obj/item/W in L)
 		if(!L.unEquip(W))
 			qdel(W)
-	/*visible_message(
-		"<span class='hierophant'>\"[pick(kill_phrases)]\"</span>\n<span class='hierophant_warning'>[src] annihilates [L]!</span>",
-		"<span class='userdanger'>You annihilate [L], restoring your health!</span>")*/
 	visible_message(
-		"<span class='hierophant'>\"Caw.\"</span>\n<span class='hierophant_warning'>[src] annihilates [L]!</span>",
+		"<span class='hierophant'>\"[pick(kill_phrases)]\"</span>\n<span class='hierophant_warning'>[src] annihilates [L]!</span>",
 		"<span class='userdanger'>You annihilate [L], restoring your health!</span>")
 	adjustHealth(-L.maxHealth*0.5)
 	L.dust()
 
-/*/mob/living/simple_animal/hostile/megafauna/hierophant/GiveTarget(new_target)
+/mob/living/simple_animal/hostile/megafauna/hierophant/GiveTarget(new_target)
 	var/targets_the_same = (new_target == target)
 	. = ..()
 	if(. && target && !targets_the_same)
-		visible_message("<span class='hierophant'>\"[pick(target_phrases)]\"</span>")*/
+		visible_message("<span class='hierophant'>\"[pick(target_phrases)]\"</span>")
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/adjustHealth(amount)
 	. = ..()
@@ -149,8 +145,8 @@ Difficulty: Hard
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/Move()
 	if(!blinking)
-		/*if(!stat)
-			playsound(loc, 'sound/mecha/mechmove04.ogg', 150, 1, -4)*/
+		if(!stat)
+			playsound(loc, 'sound/mecha/mechmove04.ogg', 150, 1, -4)
 		..()
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/Goto(target, delay, minimum_distance)
@@ -197,7 +193,7 @@ Difficulty: Hard
 			switch(pick(possibilities))
 				if("blink_spam") //blink either once or multiple times.
 					if(health < maxHealth * 0.5 && !target_is_slow && blink_counter > 1)
-						//visible_message("<span class='hierophant'>\"Mx ampp rsx iwgeti.\"</span>")
+						visible_message("<span class='hierophant'>\"Mx ampp rsx iwgeti.\"</span>")
 						var/oldcolor = color
 						animate(src, color = "#660099", time = 6)
 						while(health && target && blink_counter)
@@ -215,7 +211,7 @@ Difficulty: Hard
 					else
 						blink(target)
 				if("cross_blast_spam") //fire a lot of cross blasts at a target.
-					//visible_message("<span class='hierophant'>\"Piezi mx rsalivi xs vyr.\"</span>")
+					visible_message("<span class='hierophant'>\"Piezi mx rsalivi xs vyr.\"</span>")
 					blinking = TRUE
 					var/oldcolor = color
 					animate(src, color = "#660099", time = 6)
@@ -233,7 +229,7 @@ Difficulty: Hard
 					sleep(8)
 					blinking = FALSE
 				if("chaser_swarm") //fire four fucking chasers at a target and their friends.
-					//visible_message("<span class='hierophant'>\"Mx gerrsx lmhi.\"</span>")
+					visible_message("<span class='hierophant'>\"Mx gerrsx lmhi.\"</span>")
 					blinking = TRUE
 					var/oldcolor = color
 					animate(src, color = "#660099", time = 10)
@@ -289,8 +285,7 @@ Difficulty: Hard
 	if(!T)
 		return
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph/diagonal, list(T, src))
-	playsound(T,'sound/magic/blink.ogg', 200, 1)
-	//playsound(T,'sound/effects/bin_close.ogg', 200, 1)
+	playsound(T,'sound/effects/bin_close.ogg', 200, 1)
 	sleep(2)
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/blast, list(T, src, FALSE))
 	for(var/d in diagonals)
@@ -301,8 +296,7 @@ Difficulty: Hard
 	if(!T)
 		return
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph/cardinal, list(T, src))
-	playsound(T,'sound/magic/blink.ogg', 200, 1)
-	//playsound(T,'sound/effects/bin_close.ogg', 200, 1)
+	playsound(T,'sound/effects/bin_close.ogg', 200, 1)
 	sleep(2)
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/blast, list(T, src, FALSE))
 	for(var/d in cardinal)
@@ -313,8 +307,7 @@ Difficulty: Hard
 	if(!T)
 		return
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph, list(T, src))
-	playsound(T,'sound/magic/blink.ogg', 200, 1)
-	//playsound(T,'sound/effects/bin_close.ogg', 200, 1)
+	playsound(T,'sound/effects/bin_close.ogg', 200, 1)
 	sleep(2)
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/blast, list(T, src, FALSE))
 	for(var/d in alldirs)
@@ -336,10 +329,8 @@ Difficulty: Hard
 	var/turf/source = get_turf(src)
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph, list(T, src))
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph, list(source, src))
-	playsound(T,'sound/magic/blink.ogg', 200, 1)
-	//playsound(T,'sound/magic/Wand_Teleport.ogg', 200, 1)
-	playsound(source,'sound/magic/blink.ogg', 200, 1)
-	//playsound(source,'sound/machines/AirlockOpen.ogg', 200, 1)
+	playsound(T,'sound/magic/Wand_Teleport.ogg', 200, 1)
+	playsound(source,'sound/machines/AirlockOpen.ogg', 200, 1)
 	blinking = TRUE
 	sleep(2) //short delay before we start...
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph/teleport, list(T, src))
@@ -371,15 +362,13 @@ Difficulty: Hard
 	if(!T)
 		return
 	PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph, list(T, src))
-	playsound(T,'sound/magic/blink.ogg', 200, 1)
-	//playsound(T,'sound/effects/bin_close.ogg', 200, 1)
+	playsound(T,'sound/effects/bin_close.ogg', 200, 1)
 	sleep(2)
 	for(var/t in RANGE_TURFS(1, T))
 		PoolOrNew(/obj/effect/overlay/temp/hierophant/blast, list(t, src, FALSE))
 
 /mob/living/simple_animal/hostile/megafauna/hierophant/proc/burst(turf/original) //release a wave of blasts
-	playsound(original,'sound/magic/blink.ogg', 200, 1)
-	//playsound(original,'sound/machines/AirlockOpen.ogg', 200, 1)
+	playsound(original,'sound/machines/AirlockOpen.ogg', 200, 1)
 	var/last_dist = 0
 	for(var/t in spiral_range_turfs(burst_range, original))
 		var/turf/T = t
@@ -545,56 +534,41 @@ Difficulty: Hard
 		M.take_damage(damage, BURN, 0, 0)
 
 /obj/effect/hierophant
-	name = "hierophant rune"
-	desc = "A powerful magic mark allowing whomever attunes themself to it to return to it at will."
-	icon = 'icons/obj/rune.dmi'
-	icon_state = "hierophant"
+	name = "hierophant beacon"
+	desc = "A strange beacon, allowing mass teleportation from those able to use it."
+	icon = 'icons/obj/lavaland/artefacts.dmi'
+	icon_state = "hierophant_tele_off"
 	layer = LOW_OBJ_LAYER
 	anchored = TRUE
-	color = "#CC00FF"
 
 /obj/effect/hierophant/attackby(obj/item/I, mob/user, params)
-	if(istype(I, /obj/item/weapon/hierophant_staff))
-		var/obj/item/weapon/hierophant_staff/H = I
-		if(H.rune == src)
-			user << "<span class='notice'>You start removing your hierophant rune...</span>"
+	if(istype(I, /obj/item/weapon/hierophant_club))
+		var/obj/item/weapon/hierophant_club/H = I
+		if(H.timer > world.time)
+			return
+		if(H.beacon == src)
+			user << "<span class='notice'>You start removing your hierophant beacon...</span>"
 			H.timer = world.time + 51
+			H.prepare_icon_update()
 			if(do_after(user, 50, target = src))
 				playsound(src,'sound/magic/Blind.ogg', 200, 1, -4)
 				PoolOrNew(/obj/effect/overlay/temp/hierophant/telegraph/teleport, list(get_turf(src), user))
-				user << "<span class='hierophant_warning'>You touch the rune with the staff, dispelling it!</span>"
-				H.rune = null
+				user << "<span class='hierophant_warning'>You collect [src], reattaching it to the club!</span>"
+				H.beacon = null
 				user.update_action_buttons_icon()
 				qdel(src)
 			else
 				H.timer = world.time
+				H.prepare_icon_update()
 		else
-			user << "<span class='hierophant_warning'>You touch the rune with the staff, but nothing happens.</span>"
-
+			user << "<span class='hierophant_warning'>You touch the beacon with the club, but nothing happens.</span>"
 	else
-		..()
+		return ..()
 
 /obj/item/device/gps/internal/hierophant
 	icon_state = null
 	gpstag = "Zealous Signal"
 	desc = "Heed its words."
 	invisibility = 100
-
-/turf/open/indestructible/hierophant
-	icon_state = "hierophant1"
-	initial_gas_mix = "o2=14;n2=23;TEMP=300"
-	desc = "A floor with a square pattern. It's faintly cool to the touch."
-
-/turf/open/indestructible/hierophant/New()
-	..()
-	if(prob(50))
-		icon_state = "hierophant2"
-
-/turf/closed/indestructible/riveted/hierophant
-	name = "wall"
-	desc = "A wall made out of smooth, cold stone."
-	icon = 'icons/turf/walls/hierophant_wall.dmi'
-	icon_state = "hierophant"
-	smooth = SMOOTH_TRUE
 
 #undef MEDAL_PREFIX
